Revert "refactored: decoupled downloader and preview helper/util functions"
This reverts commit 95d63735af77ea972bdc8b74227c134534d511b2.
